Fix critical issues: decimal precision and silent exceptions
1. Purchase model: Float → Numeric(10,2) for amount/total columns
   - Prevents floating-point precision errors in financial calculations
   - Matches Tip model which already uses Decimal(10,2)
   - Migration 0021 handles PostgreSQL column type conversion

2. Replace 5 bare `except:` clauses with proper logging
   - All were in wallet balance fetch for error pages
   - Now logs warning with actual error message
   - Makes debugging possible when wallet queries fail

+++ models.py
@@ -302,8 +302,8 @@ class Purchase(Base):
     artist_id = Column(String(50), nullable=True, index=True)
     item_id = Column(String(50), nullable=True, index=True)
     qty = Column(Integer, nullable=False, default=1)
-    amount = Column(Float, nullable=False)         # amount before fees
-    total = Column(Float, nullable=False)          # final charge
+    amount = Column(Numeric(10, 2), nullable=False)         # amount before fees
+    total = Column(Numeric(10, 2), nullable=False)          # final charge
     stripe_id = Column(String(255), nullable=True, index=True)
     status = Column(String(50), nullable=False, default="pending", index=True)
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False, index=True)
